Moved the test into BugTest.scala from JDBCConnectionPoolTestSuite
